+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Changed
 
 - Removed greedy matching of dimension order in Bio-Formats Zarr output. Just check if OME-Zarr.
+- Use WASM LZW decoder from `manzt/geotiff.js`
 
 ## 0.4.2