+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Added
 
+- Support local tiff files.
+- Support local zarr directories.
+
 ### Changed
 
 - Removed greedy matching of dimension order in Bio-Formats Zarr output. Just check if OME-Zarr.